analysis/requirements.txt
pandas
python-decouple
unidecode
urlib2
httpx